# backend/Models/openAPI.yml
openapi: 3.0.3
info:
  title: Brewtiful API
  version: 1.0.0
servers:
  - url: http://{{baseurl}}
...
paths:
...
  /api/Item:
    get:
      summary: Get all items
      deprecated: false
      description: ''
      tags:
        - api/Item
      parameters:
        - name: Accept
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            '*/*':
              schema:
                type: object
                properties: {}
              examples:
                '1':
                  summary: Get all items
                  value: "{\r\n    \"id\": \"integer\",\r\n    \"name\": \"string\",\r\n    \"categoryId\": \"integer\",\r\n    \"category\": {\r\n        \"id\": \"integer\",\r\n        \"name\": \"string\"\r\n    },\r\n    \"ingredientNames\": \"array<string>\",\r\n    \"ingredients\": [\r\n        {\r\n            \"id\": \"integer\",\r\n            \"name\": \"string\",\r\n            \"defaultQuantity\": \"integer\",\r\n            \"extraCost\": \"float\"\r\n        },\r\n        {\r\n            \"id\": \"integer\",\r\n            \"name\": \"string\",\r\n            \"defaultQuantity\": \"integer\",\r\n            \"extraCost\": \"float\"\r\n        }\r\n    ],\r\n    \"price\": \"float\"\r\n}\r\n\r\n{\r\n    \"id\": \"integer\",\r\n    \"name\": \"string\",\r\n    \"categoryId\": \"integer\",\r\n    \"category\": {\r\n        \"id\": \"integer\",\r\n        \"name\": \"string\"\r\n    },\r\n    \"ingredientNames\": \"array<string>\",\r\n    \"ingredients\": [\r\n        {\r\n            \"id\": \"integer\",\r\n            \"name\": \"string\",\r\n            \"defaultQuantity\": \"integer\",\r\n            \"extraCost\": \"float\"\r\n        },\r\n        {\r\n            \"id\": \"integer\",\r\n            \"name\": \"string\",\r\n            \"defaultQuantity\": \"integer\",\r\n            \"extraCost\": \"float\"\r\n        }\r\n    ],\r\n    \"price\": \"float\"\r\n}\r\n"
          headers: {}
      security: []
    post:
      summary: Invalid fields (400)
      deprecated: false
      description: ''
      tags:
        - api/Item
      parameters:
        - name: Content-Type
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
        - name: Accept
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Id:
                  type: integer
                Name:
                  type: string
                CategoryId:
                  type: integer
                Category:
                  type: object
                  properties:
                    Id:
                      type: integer
                    Name:
                      type: string
                  required:
                    - Id
                    - Name
                IngredientNames:
                  type: array
                  items:
                    type: string
                Ingredients:
                  type: array
                  items:
                    type: object
                    properties:
                      Id:
                        type: integer
                      Name:
                        type: string
                      DefaultQuantity:
                        type: integer
                      ExtraCost:
                        type: integer
                    required:
                      - Id
                      - Name
                      - DefaultQuantity
                      - ExtraCost
                Price:
                  type: integer
              required:
                - Id
                - Name
                - CategoryId
                - Category
                - IngredientNames
                - Ingredients
                - Price
            example:
              Id: 10
              Name: Latte Updated2
              CategoryId: 1
              Category:
                Id: 1
                Name: Kava
              IngredientNames:
                - Pienas
                - Espresso
              Ingredients:
                - Id: 1
                  Name: Pienas
                  DefaultQuantity: 1
                  ExtraCost: 0
                - Id: 2
                  Name: Espresso
                  DefaultQuantity: 1
                  ExtraCost: 1
              Price: 5
      responses:
        '400':
          description: ''
          content:
            '*/*':
              schema:
                type: object
                properties:
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                required:
                  - title
                  - status
                  - detail
              examples:
                '1':
                  summary: Invalid fields (400)
                  value: "{\r\n    \"title\": \"BadRequest\",\r\n    \"status\": 400,\r\n    \"detail\": \"Item name can only contain English and Lithuanian letters.\"\r\n}"
          headers: {}
      security: []
...